Fixed merge conflicts cuz of EventDetailsEntity

+++ api/MorganizerAPI/src/main/java/com/morganizer/entity/TaskListItemsEntity.java
@@ -49,7 +49,7 @@ public class TaskListItemsEntity {
 		this.task = task;
 	}
 
-	public int getItemId() {
+	public long getItemId() {
 		return itemId;
 	}
